requirements.txt
requests
pyshout
pytz
python-dotenv
mutagen
openai